app/requirements.txt
uvicorn==0.30.1
fastapi==0.111.0
stripe==10.0.0
fastapi-jwt-auth==0.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.31
pymysql==1.1.1
alembic==1.13.1